+++ typed_sql/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.1.3-wip
  * Added `TransactionAbortedException.toString()` to render the `reason` when
    printing an exception.
+ * Fix postgres support for `DateTime.utc(0)`, which must be encoded as 'BC'
+   suffixed date-time.
 
 ## 0.1.2
  * Support for _composite foreign keys_ using the `@ForeignKey` annotation.